requirements.txt
scikit-learn
numpy
scipy
SPLAT-library
nltk
xml2json
